fix(calendar): add story for views in calendar component
