+++ core/components/organisms/calendar/__stories__/variants/size.story.jsx
@@ -5,33 +5,28 @@ import Heading from '@/components/atoms/heading';
 
 // CSF format story
 export const size = () => {
-  const view = ['year', 'month', 'date'];
 
   return (
     <>
       <div className="mt-8">
         <Heading>size: small</Heading>
         <div className="d-flex">
-          {view.map((v, index) => (
-            <div className="mr-8" key={index}>
-              <Card className="d-inline-flex" shadow="light">
-                <Calendar date={new Date(2020, 2, 15)} size={'small'} view={v} />
-              </Card>
-            </div>
-          ))}
+          <div className="mr-8">
+            <Card className="d-inline-flex" shadow="light">
+              <Calendar date={new Date(2020, 2, 15)} size={'small'} view='date' />
+            </Card>
+          </div>
         </div>
       </div>
 
       <div className="mt-8">
         <Heading>size: large</Heading>
         <div className="d-flex">
-          {view.map((v, index) => (
-            <div className="mr-8" key={index}>
-              <Card className="d-inline-flex" shadow="light">
-                <Calendar date={new Date(2020, 2, 15)} size={'large'} view={v} />
-              </Card>
-            </div>
-          ))}
+          <div className="mr-8">
+            <Card className="d-inline-flex" shadow="light">
+              <Calendar date={new Date(2020, 2, 15)} size={'large'} view='date' />
+            </Card>
+          </div>
         </div>
       </div>
     </>
